apps/api/pyproject.toml
[project]
name = "color-analysis-api"
version = "0.1.0"
description = "Seasonal color analysis API and worker"
requires-python = ">=3.12,<3.14"
dependencies = [
  "alembic==1.14.1",
  "boto3==1.36.2",
  "fastapi==0.115.8",
  "httpx==0.28.1",
  "mediapipe==0.10.33",
  "numpy==2.2.2",
  "opencv-python-headless==4.11.0.86",
  "pillow==11.1.0",
  "pillow-heif==0.21.0",
  "pydantic-settings==2.7.1",
  "python-multipart==0.0.20",
  "redis==5.2.1",
  "rq==2.1.0",
  "sqlalchemy==2.0.37",
  "asyncpg==0.30.0",
  "greenlet==3.1.1",
  "uvicorn==0.34.0"
]

[project.optional-dependencies]
dev = [
  "mypy==1.14.1",
  "pytest==8.3.4",
  "pytest-asyncio==0.25.3",
  "pytest-cov==6.0.0",
  "ruff==0.9.2",
  "syrupy==4.8.1"
]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008", "E501", "I001", "UP017"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_unused_configs = true
plugins = []
mypy_path = "src"
disable_error_code = ["arg-type", "assignment", "union-attr", "call-overload", "misc"]

[[tool.mypy.overrides]]
module = ["mediapipe.*", "cv2", "pillow_heif"]
ignore_missing_imports = true
